requirements.txt
numpy
pandas
prettytable
tensorflow
scikit-image
matplotlib
scikit-learn
pyyaml
wandb
